Use an all-js sync IPC mechanism using worker threads + an http server
This is so outrageous and I apologize!

I was having several local issues with the node-sync-ipc package including segfaults and other strange error conditions that just don't happen when you use JS modules. In order to remove that dependency and hopefully increase stability, there's a hack that a contributor to esbuild suggested to use Atomics.wait in the main thread to block it when you need to respond to `require` synchronously, and then do the async stuff you need to do in a worker thread, which then wakes the main thread up. Lets see if that is more stable.