Set the default sourcemap type to inline
esbuild-dev currently hooks into the node require process at a kind of sneaky point where it serves compiled JS as the file contents of TS files. The requiring node process doesn't really know much about the temporary directory esbuild is building ito, it's the require extension that ninjas in the real JS content of the file.

For this reason, a relative sourcemap url inside that served JS is interpreted by node and by vscode as relative to the actual user's working directory, not the esbuild-dev workdir. This is a problem because an external sourcemap is stuck in the workdir, not the user's source root! I couldn't find an option in esbuild to control where the sourcemaps are written out to, so, lets just switch the sourcemap to being inline inside the compiled JS so node / vscode can always find it. Woop woop!