Drop buggy resiliency code to node bug now that node bug has been fixed
When wds first came into existence, node's `worker_threads` module had a bug where sometimes a worker thread's data wouldn't be synchronously available to it. This made the `SyncWorker` code need to be careful -- if the bug had triggered, it needed to wait for a bit until the data was available to start working. This `setImmediate` call worked for this purpose, but, was an infinite loop bug lying in wait!

`SyncWorker` has the unfortunate privilege of always being required by any code running under wds, in the main thread or in any other worker threads. Node passes the same `--require` flags to worker threads, so they get the same registration stuff that wds sets up, which is desirable so that worker threads can also require TypeScript code using wds. This makes `SyncWorker` need to be smart though, and be aware of if it is being required by a "parent" thread, be it a main thread or any not-wds worker thread, or if it is being required by a "child" thread for powering the parent's SyncWorker. The `SyncWorker` code needs the worker data for this, and without it, doesn't know what to do.

If no worker data is passed to some non-wds worker thread, SyncWorker will be required, but not see any worker data, and trigger this `setImmediate` defense mechanism infinitely. Oops. We can't know if the worker data is absent cause of the bug, or if it is absent cause it really wasn't passed and it is just not a `SyncWorker` thread. So instead, we choose to only support node versions without the bug. Kinda unfortunate, but also, not really sure there is another option.