Remove 'supervise' mode in favor of 'watch' mode
There's no point in keeping the process alive on 'watch' mode, but
being unable to reload on any file changes. Before this fix, we would
need to use 'watch' and 'supervise' modes together to achieve this.

This fix makes 'watch' mode behave like 'supervise' mode, but getting
rid of the latter.