Kill the whole process group when shutting down, instead of just the one child process
This fixes a bug where when wds is orchestrating a process that starts
its own child processes, those child processes can be come zombies if we
just SIGTERM/SIGKILL the main orchestrated process. What we need to do
is kill the whole process group, which you can do by sending the
negative pid of the parent to the kernel.

See https://azimi.me/2014/12/31/kill-child_process-node-js.html