Write files to disk atomically so they are never empty
This fixes a bug where if multiple parallel worker threads in a wds project start around the same time, one will trigger a build of the whole project and another will soon after trigger a build of just one file. Both builds are processed in parallel for speed, but if the timing is just right, both try to write to the same destination file around the same time. If one of the project threads is unlucky enough to read the file at around when it is being written, it can read an empty or half written file. Whoops!

`write-file-atomic` solves this problem by using a tmpfile and serializing writes made from within the same process.

COMPUTERS

[no-changelog-required]